+++ NEWS.md
@@ -1,3 +1,5 @@
++ Added Anderson-Darling normality testing `anderson.darling.testDS` to perform normality testing on samples with over 5000 individuals.
+
 # dsExposome 2.0.2
 
 + Changed output of `invExWASDS` function to return a list of tables instead of an `ExposomeSet` object.